refactor(agent): extract session title generator
