+++ electron/agent/manager.ts
@@ -10,7 +10,7 @@ import type {
 import type { ModelChoice } from "../models/common.ts"
 import type { PersistedCustomModel } from "../models/store.ts"
 import type { GenerateSessionTitleRequest, SessionInfo } from "../session/common.ts"
-import type { BuildSessionTitleInput } from "../session/title.ts"
+import type { GeneratedSessionTitle } from "./session-title-generator.ts"
 import type { FilePartInput, SessionPromptAsyncData, TextPartInput } from "@opencode-ai/sdk/v2/client"
 import type { OpencodeClient } from "@opencode-ai/sdk/v2/client"
 
@@ -23,7 +23,6 @@ import { branding } from "../branding.ts"
 import { logDiagnostic } from "../diagnostics-log.ts"
 import { connectorBaseUrl, llmBaseUrl } from "../domain.ts"
 import { DEFAULT_BUILTIN_MODEL_ID, isBuiltinModelId, resolveBuiltinModel } from "../models/builtin.ts"
-import { buildFallbackSessionTitle, sanitizeGeneratedSessionTitle } from "../session/title.ts"
 import { buildOpencodeConfig, customProviderId, WANTA_MODEL_ID, WANTA_PROVIDER_ID } from "./config.ts"
 import { normalizeMessage, normalizePermissionRequest, normalizeQuestionRequest } from "./event-translator.ts"
 import { normalizeWantaAgentMode } from "./mode.ts"
@@ -31,8 +30,11 @@ import { writeOoIdentitySettings } from "./oo-identity.ts"
 import { buildOoEnv } from "./oo.ts"
 import { managedPythonEnvironmentPath, managedPythonExecutable } from "./python-environment.ts"
 import { opencodeReasoningVariant } from "./reasoning.ts"
+import { generateSessionTitle as generateTitle } from "./session-title-generator.ts"
 import { OpencodeSidecar } from "./sidecar.ts"
 import { ensureAgentWorkspace } from "./workspace.ts"
+
+export type { GeneratedSessionTitle } from "./session-title-generator.ts"
 
 export interface AgentManagerOptions {
   /** 网关鉴权凭证：现为会话 token（网关层接受 cookie/token/api-key）。LLM 网关 / connector / oo-cli 共用。 */
@@ -122,31 +124,6 @@ export interface RawSession {
   time?: { created?: number; updated?: number }
 }
 
-interface ChatCompletionResponse {
-  choices?: Array<{ message?: { content?: string } }>
-}
-
-export interface GeneratedSessionTitle {
-  title: string
-  generated: boolean
-}
-
-const sessionTitleSystemPrompt = [
-  "Generate a concise chat title as a task label.",
-  'Return JSON only, exactly like {"title":"Gmail 三日报告"}.',
-  "Return the JSON immediately without analysis or reasoning.",
-  "Keep the user's language when possible.",
-  "Keep Chinese titles within about 10 characters and English titles within 6 words.",
-  "- Preserve complete brand, product, app, domain, and file names. Never cut Gmail to Gma or truncate any word.",
-  "- Prefer the core action and object; remove polite wording such as help me, 请, 帮我, 麻烦.",
-  "- No URLs, no ellipses, no markdown, no explanations, no trailing punctuation.",
-  "Examples:",
-  'User: 你帮我分析一下我最近三天的 Gmail 信息，然后给我总结出一个报告 -> {"title":"Gmail 三日报告"}',
-  'User: 你帮我将这个店铺中商品相关的图片都抓下来 -> {"title":"抓取店铺商品图片"}',
-  'User: Search 1688 product images with Metaso and Puppeteer -> {"title":"1688 Product Images"}',
-].join("\n")
-const sessionTitleRequestTimeoutMs = 30_000
-const sessionTitleMaxTokens = 512
 const eventStreamMaxReconnectAttempts = 5
 const eventStreamRestartInitialDelayMs = 500
 const eventStreamRestartMaxDelayMs = 5_000
@@ -590,65 +567,9 @@ export class AgentManager {
     await this.client.session.delete({ sessionID: id })
   }
 
-  public async generateSessionTitle(input: GenerateSessionTitleRequest): Promise<GeneratedSessionTitle> {
-    const fallback = buildFallbackSessionTitle(input)
-    const titleSource = buildTitleSource(input)
-    if (!titleSource) {
-      return { generated: false, title: fallback }
-    }
-
-    try {
-      const rawTitle = await this.requestSessionTitle(titleSource, input.model)
-      const title = sanitizeGeneratedSessionTitle(rawTitle, input)
-      return title.usedFallback ? { generated: false, title: fallback } : { generated: true, title: title.title }
-    } catch (error) {
-      console.warn("[wanta] failed to generate session title, using fallback:", error)
-      return { generated: false, title: fallback }
-    }
+  public generateSessionTitle(input: GenerateSessionTitleRequest): Promise<GeneratedSessionTitle> {
+    return generateTitle(input, (choice) => this.resolveSessionTitleTarget(choice))
   }
-
-  private async requestSessionTitle(
-    titleSource: string,
-    modelChoice: ModelChoice | undefined,
-    previousTitle?: string,
-  ): Promise<string> {
-    const target = this.resolveSessionTitleTarget(modelChoice)
-    const messages = [
-      {
-        role: "system",
-        content: sessionTitleSystemPrompt,
-      },
-      {
-        role: "user",
-        content: previousTitle
-          ? `Rewrite the title because it violated the length or quality rules: ${JSON.stringify(previousTitle)}\n\nSource:\n${titleSource}`
-          : titleSource,
-      },
-    ]
-    const headers: Record<string, string> = {
-      "Content-Type": "application/json",
-    }
-    if (target.apiKey) {
-      headers.Authorization = `Bearer ${target.apiKey}`
-    }
-    const response = await fetch(`${target.baseUrl.replace(/\/+$/, "")}/chat/completions`, {
-      method: "POST",
-      headers,
-      body: JSON.stringify({
-        model: target.modelID,
-        temperature: 0.1,
-        max_tokens: sessionTitleMaxTokens,
-        messages,
-      }),
-      signal: AbortSignal.timeout(sessionTitleRequestTimeoutMs),
-    })
-    if (!response.ok) {
-      throw new Error(`session title request failed: ${response.status} ${response.statusText}`)
-    }
-    const payload = (await response.json()) as ChatCompletionResponse
-    return payload.choices?.[0]?.message?.content ?? ""
-  }
-
   private resolveSessionTitleTarget(choice: ModelChoice | undefined): {
     apiKey: string
     baseUrl: string
@@ -1110,13 +1031,6 @@ function mergeSystemPrompts(...parts: Array<string | undefined>): string | undef
     .filter((part): part is string => Boolean(part))
     .join("\n\n")
   return merged || undefined
-}
-
-function buildTitleSource(input: BuildSessionTitleInput): string {
-  const parts = [input.text, ...(input.attachmentNames ?? []).map((name) => `Attachment: ${name}`)]
-    .map((part) => part.trim())
-    .filter(Boolean)
-  return parts.join("\n").slice(0, 1600)
 }
 
 function pathToFileUrl(filePath: string): string {
